+++ es-core/src/math/Transform4x4f.cpp
@@ -138,8 +138,8 @@ Transform4x4f& Transform4x4f::rotate(const float _angle, const Vector3f& _axis)
 {
 	float*       tm      = (float*)this;
 	const float* av      = (float*)&_axis;
-	const float  s       = Math::sin(-_angle);
-	const float  c       = Math::cos(-_angle);
+	const float  s       = Math::sinf(-_angle);
+	const float  c       = Math::cosf(-_angle);
 	const float  t       = 1 - c;
 	const float  x       = av[0];
 	const float  y       = av[1];
@@ -186,8 +186,8 @@ Transform4x4f& Transform4x4f::rotate(const float _angle, const Vector3f& _axis)
 Transform4x4f& Transform4x4f::rotateX(const float _angle)
 {
 	float*      tm      = (float*)this;
-	const float s       = Math::sin(-_angle);
-	const float c       = Math::cos(-_angle);
+	const float s       = Math::sinf(-_angle);
+	const float c       = Math::cosf(-_angle);
 	const float temp[6] = { tm[ 1] *  c + tm[ 2] * s,
 		                    tm[ 1] * -s + tm[ 2] * c,
 		                    tm[ 5] *  c + tm[ 6] * s,
@@ -209,8 +209,8 @@ Transform4x4f& Transform4x4f::rotateX(const float _angle)
 Transform4x4f& Transform4x4f::rotateY(const float _angle)
 {
 	float*      tm      = (float*)this;
-	const float s       = Math::sin(-_angle);
-	const float c       = Math::cos(-_angle);
+	const float s       = Math::sinf(-_angle);
+	const float c       = Math::cosf(-_angle);
 	const float temp[6] = { tm[ 0] * c + tm[ 2] * -s,
 		                    tm[ 0] * s + tm[ 2] *  c,
 		                    tm[ 4] * c + tm[ 6] * -s,
@@ -232,8 +232,8 @@ Transform4x4f& Transform4x4f::rotateY(const float _angle)
 Transform4x4f& Transform4x4f::rotateZ(const float _angle)
 {
 	float*      tm      = (float*)this;
-	const float s       = Math::sin(-_angle);
-	const float c       = Math::cos(-_angle);
+	const float s       = Math::sinf(-_angle);
+	const float c       = Math::cosf(-_angle);
 	const float temp[6] = { tm[ 0] *  c + tm[ 1] * s,
 		                    tm[ 0] * -s + tm[ 1] * c,
 		                    tm[ 4] *  c + tm[ 5] * s,
@@ -269,9 +269,9 @@ Transform4x4f& Transform4x4f::round()
 {
 	float* tm = (float*)this;
 
-	tm[12] = (int)(tm[12] + 0.5f);
-	tm[13] = (int)(tm[13] + 0.5f);
-	tm[14] = (int)(tm[14] + 0.5f);
+	tm[12] = (float)(int)(tm[12] + 0.5f);
+	tm[13] = (float)(int)(tm[13] + 0.5f);
+	tm[14] = (float)(int)(tm[14] + 0.5f);
 
 	return *this;
