+++ es-core/src/math/Vector2f.cpp
@@ -2,8 +2,8 @@
 
 Vector2f& Vector2f::round()
 {
-	mX = (int)(mX + 0.5f);
-	mY = (int)(mY + 0.5f);
+	mX = (float)(int)(mX + 0.5f);
+	mY = (float)(int)(mY + 0.5f);
 		
 	return *this;
